Rewrite of the parser
The previous version had a big limitation, which was we needed to have only one
regexp for each caracteristic. To overcome this the developer would have to master
regular expressions and their complexity.

Now the parser contains a list of regexp for each parameter. The regexps will run
in order. This is to remove well known terms from the name to reduce the more
expressive parameters (the ones that can englobe everything).